Reduce BeamFitter memory footprint: store BSTrkParameters in float instead of double

+++ RecoVertex/BeamSpotProducer/interface/BSTrkParameters.h
@@ -5,11 +5,7 @@
    class:   BSTrkParameters.h
    package: RecoVertex/BeamSpotProducer
    
-
-
- author: Francisco Yumiceva, Fermilab (yumiceva@fnal.gov)
-
-
+   author: Francisco Yumiceva, Fermilab (yumiceva@fnal.gov)
 ________________________________________________________________**/
 
 class BSTrkParameters {
@@ -17,26 +13,14 @@ public:
   // constructor
   BSTrkParameters() {}
   // constructor from values
-  //BSTrkParameters( double z0, double sigz0,
-  //				 double d0, double sigd0,
-  //				 double phi0, double pt) {
-  //	fz0 = z0;
-  //	fsigz0 = sigz0;
-  //	fd0 = d0;
-  //	fsigd0 = sigd0;
-  //	fphi0 = phi0;
-  //	fpt = pt;
-  //
-  //};
-
-  BSTrkParameters(double z0,
-                  double sigz0,
-                  double d0,
-                  double sigd0,
-                  double phi0,
-                  double pt,
-                  double d0phi_d0 = 0.,
-                  double d0phi_chi2 = 0.) {
+  BSTrkParameters(float z0,
+                  float sigz0,
+                  float d0,
+                  float sigd0,
+                  float phi0,
+                  float pt,
+                  float d0phi_d0 = 0.,
+                  float d0phi_chi2 = 0.) {
     fz0 = z0;
     fsigz0 = sigz0;
     fd0 = d0;
@@ -50,30 +34,30 @@ public:
   };
 
   //
-  double z0() const { return fz0; }
-  double sigz0() const { return fsigz0; }
-  double d0() const { return fd0; }
-  double sigd0() const { return fsigd0; }
-  double phi0() const { return fphi0; }
-  double pt() const { return fpt; }
-  double d0phi_chi2() const { return fd0phi_chi2; }
-  double d0phi_d0() const { return fd0phi_d0; }
-  double vx() const { return fvx; }
-  double vy() const { return fvy; }
-  void setVx(double vx) { fvx = vx; }
-  void setVy(double vy) { fvy = vy; }
+  float z0() const { return fz0; }
+  float sigz0() const { return fsigz0; }
+  float d0() const { return fd0; }
+  float sigd0() const { return fsigd0; }
+  float phi0() const { return fphi0; }
+  float pt() const { return fpt; }
+  float d0phi_chi2() const { return fd0phi_chi2; }
+  float d0phi_d0() const { return fd0phi_d0; }
+  float vx() const { return fvx; }
+  float vy() const { return fvy; }
+  void setVx(float vx) { fvx = vx; }
+  void setVy(float vy) { fvy = vy; }
 
 private:
-  double fz0;
-  double fsigz0;
-  double fd0;
-  double fsigd0;
-  double fphi0;
-  double fpt;
-  double fd0phi_chi2;
-  double fd0phi_d0;
-  double fvx;
-  double fvy;
+  float fz0;
+  float fsigz0;
+  float fd0;
+  float fsigd0;
+  float fphi0;
+  float fpt;
+  float fd0phi_chi2;
+  float fd0phi_d0;
+  float fvx;
+  float fvy;
 };
 
 #endif
